Making the MPLAB project compile
