+++ firmware/kalman_board.X/kalman_lib/kalman_test.c
@@ -1,3 +1,4 @@
+/*
 #include "data.h"
 #include "orientation_conversion.h"
 #include <stdio.h>
@@ -25,9 +26,9 @@ int main() {
 
         printf("%f\n", get_orientation());
         // I take it that Angle States are x_sl and you meant to use get_orientation
-        struct Matrix Conv = reference_frame_correction(vel, get_orientation(), Conv.data);
+        //struct Matrix Conv = reference_frame_correction(vel, get_orientation(), Conv.data);
             
-        struct Vector A_corr = vector_multiplication(Conv, (struct Vector) {a_prev, 3} , a_corr);
+        //struct Vector A_corr = vector_multiplication(Conv, (struct Vector) {a_prev, 3} , a_corr);
         //struct Vector X_corr = vector_multiplication(Conv, (struct Vector) {x_prev, 3} , x_corr);
 
         update_velocity_filter(new_time, x_prev[0], a_corr[0], x_prev[1], a_corr[1], x_prev[2], a_corr[2]);
@@ -38,3 +39,4 @@ int main() {
     }
     
 }
+ */
